refactor(desktop): move build_gui_app into uc-desktop, slim main.rs to 12 lines
uc-bootstrap now stays GUI-shell agnostic: it provides composition-root
assembly tools (wire_dependencies, init_tracing_subscriber,
get_storage_paths, install_panic_logging_hook), plus CLI / daemon
entry-point builders. The GUI shell entry-point builder
(build_gui_app + GuiBootstrapContext) moves into uc-desktop, the
right home now that uc-desktop is the GUI-framework agnostic host
layer that any desktop GUI shell (uc-tauri today, uc-macos-native
planned) consumes.

uc_tauri::run additionally takes ownership of build_gui_app: the bin
no longer constructs the bootstrap context itself, only the
tauri::Context (which the generate_context! macro forces into the
bin crate). Net result: src-tauri/src/main.rs is 12 lines.
