+++ src-tauri/crates/uc-bootstrap/src/lib.rs
@@ -25,8 +25,8 @@ pub use assembly::{
 };
 pub use background_tasks::{spawn_blob_processing_tasks, BlobProcessingPorts};
 pub use builders::{
-    build_cli_context, build_cli_context_with_profile, build_daemon_app, build_gui_app,
-    build_slice1_cli_context, CliBootstrapContext, DaemonBootstrapContext, GuiBootstrapContext,
+    build_cli_context, build_cli_context_with_profile, build_daemon_app, build_slice1_cli_context,
+    CliBootstrapContext, DaemonBootstrapContext,
 };
 pub use config::load_config;
 pub use init::{
