+++ src-tauri/crates/uc-desktop/src/lib.rs
@@ -14,6 +14,7 @@ pub const DAEMON_VERSION: &str = env!("CARGO_PKG_VERSION");
 pub use uc_daemon_contract::DAEMON_API_REVISION;
 
 pub mod background;
+pub mod bootstrap;
 pub mod daemon;
 pub mod daemon_probe;
 pub mod runtime;
