+++ src-tauri/src/main.rs
@@ -2,16 +2,11 @@
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
 fn main() {
-    // Tracing and config are handled inside build_gui_app()
-    let ctx = match uc_bootstrap::build_gui_app() {
-        Ok(ctx) => ctx,
-        Err(e) => {
-            eprintln!("Bootstrap failed: {}", e);
-            std::process::exit(1);
-        }
-    };
-
     // `tauri::generate_context!()` 必须在 bin crate 调用——它读取 bin
-    // 的 Cargo.toml 同目录下的 tauri.conf.json。
-    uc_tauri::run(ctx, tauri::generate_context!());
+    // 的 Cargo.toml 同目录下的 tauri.conf.json。其它装配 + Tauri 事件
+    // 循环全部交给 `uc_tauri::run`。
+    if let Err(e) = uc_tauri::run(tauri::generate_context!()) {
+        eprintln!("Tauri shell failed: {}", e);
+        std::process::exit(1);
+    }
 }
